refactor(gateway): extract startup secret surface helpers to own module
